% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{your_object}
\alias{your_object}
\title{A vector with three strings}
\format{
A vector with three strings:
\describe{
  \item{red}{A string, it's red.}
  \item{green}{A string, it's green. Street talk for money.}
  \item{blue}{A string, really didn't see that coming. It's blue this time!}
}
}
\source{
\url{http://www.themindofsebastian.com}
}
\usage{
your_object
}
\description{
A dataset containing three strings usually linked to the
   colors of pixels on a screen.
}
\keyword{datasets}
